Use ProxyPassword from the resource definition
The proxy password is defined at the resource level, and
we do not redifine it as a property before trying to update it.

So do like for languages, check if it changed and set it separately.

Also fix the configure recipe as the proxy_password was never given,
so we would have never updated it.