Use path to locate powershell.exe & remove deps on Windows cookbook
The only reason we use to have a dependency on Windows cookbook was the
locate_sysnative_cmd helper.
However this helper is not a "which" or "where", it just check whether
the given path exist under the "native" system path.
In our case, powershell.exe is not located there, but deeper in the
directory structure, so "powershell.exe" was returned.
So we don't need this helper and the Windows cookbook :)